Displacement fields are signed

+++ src/operand/address_indirect_displacement_operand.cpp
@@ -2,7 +2,7 @@
 #include "operand_validation.h"
 
 AddressIndirectDisplacementOperand::AddressIndirectDisplacementOperand(const uint8_t register_number,
-                                                                       const uint16_t displacement) :
+                                                                       const int16_t displacement) :
         register_number(operand_validation::validate_address_register_number(register_number)),
         displacement(displacement) {
 }
@@ -26,5 +26,5 @@ uint16_t AddressIndirectDisplacementOperand::get_extension_word(const uint8_t id
     if (idx != 0) {
         throw std::invalid_argument("address indirect with displacement operands have only one extension word");
     }
-    return displacement;
+    return static_cast<uint16_t>(displacement);
 }
